+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.0.0-alpha.1
 
 - **Changed:**
+  - Display: Adjust word wrap break points to have better flow
   - Update electron 19.1.5 to 19.1.6
 
 ## 1.0.0-alpha